Removing BigtableGrpcClientOption and ChannelOptions. BigtableGrpcClientOption was merged into RetryOptions and ChannelOptions was merged into BigtableOptsions.
