+++ bigtable-grpc-interface/src/main/java/com/google/cloud/bigtable/config/BigtableOptions.java
@@ -19,7 +19,8 @@ import java.net.InetAddress;
 
 import com.google.api.client.repackaged.com.google.common.annotations.VisibleForTesting;
 import com.google.api.client.util.Strings;
-import com.google.cloud.bigtable.grpc.ChannelOptions;
+import com.google.auth.Credentials;
+import com.google.cloud.bigtable.grpc.RetryOptions;
 import com.google.cloud.bigtable.grpc.TransportOptions;
 import com.google.common.base.Preconditions;
 
@@ -48,7 +49,14 @@ public class BigtableOptions {
     private InetAddress tableAdminHost;
     private InetAddress clusterAdminHost;
     private int port;
-    private ChannelOptions channelOptions;
+    private Credentials credential;
+    private String authority;
+    private String userAgent;
+    private String callStatusReportPath;
+    private String callTimingReportPath;
+    private RetryOptions retryOptions = null;
+    private long timeoutMs = 0;
+    private int channelCount = 1;
 
     public Builder setTableAdminHost(InetAddress tableAdminHost) {
       this.tableAdminHost = tableAdminHost;
@@ -85,8 +93,43 @@ public class BigtableOptions {
       return this;
     }
 
-    public Builder setChannelOptions(ChannelOptions channelOptions) {
-      this.channelOptions = channelOptions;
+    public Builder setCredential(Credentials credential) {
+      this.credential = credential;
+      return this;
+    }
+
+    public Builder setAuthority(String authority) {
+      this.authority = authority;
+      return this;
+    }
+
+    public Builder setUserAgent(String userAgent) {
+      this.userAgent = userAgent;
+      return this;
+    }
+
+    public Builder setCallStatusReportPath(String callStatusReportPath) {
+      this.callStatusReportPath = callStatusReportPath;
+      return this;
+    }
+
+    public Builder setCallTimingReportPath(String callTimingReportPath) {
+      this.callTimingReportPath = callTimingReportPath;
+      return this;
+    }
+
+    public Builder setChannelCount(int channelCount) {
+      this.channelCount  = channelCount;
+      return this;
+    }
+
+    public Builder setRetryOptions(RetryOptions retryOptions) {
+      this.retryOptions = retryOptions;
+      return this;
+    }
+
+    public Builder setTimeoutMs(long timeoutMs) {
+      this.timeoutMs = timeoutMs;
       return this;
     }
 
@@ -99,7 +142,14 @@ public class BigtableOptions {
           projectId,
           zone,
           cluster,
-          channelOptions);
+          credential,
+          authority,
+          userAgent,
+          callTimingReportPath,
+          callStatusReportPath,
+          retryOptions,
+          timeoutMs,
+          channelCount);
     }
   }
 
@@ -110,7 +160,14 @@ public class BigtableOptions {
   private final String projectId;
   private final String zone;
   private final String cluster;
-  private final ChannelOptions channelOptions;
+  private final Credentials credential;
+  private final String authority;
+  private final String userAgent;
+  private final String callTimingReportPath;
+  private final String callStatusReportPath;
+  private final RetryOptions retryOptions;
+  private final long timeoutMs;
+  private final int channelCount;
 
   @VisibleForTesting
   BigtableOptions() {
@@ -121,7 +178,14 @@ public class BigtableOptions {
       projectId = null;
       zone = null;
       cluster = null;
-      channelOptions = null;
+      credential = null;
+      authority = null;
+      userAgent = null;
+      callTimingReportPath = null;
+      callStatusReportPath = null;
+      retryOptions = null;
+      timeoutMs = 0;
+      channelCount = 1;
   }
 
   private BigtableOptions(
@@ -132,14 +196,26 @@ public class BigtableOptions {
       String projectId,
       String zone,
       String cluster,
-      ChannelOptions channelOptions) {
+      Credentials credential,
+      String authority,
+      String userAgent,
+      String callTimingReportPath,
+      String callStatusReportPath,
+      RetryOptions retryOptions,
+      long timeoutMs,
+      int channelCount) {
     Preconditions.checkArgument(
         !Strings.isNullOrEmpty(projectId), "ProjectId must not be empty or null.");
     Preconditions.checkArgument(
         !Strings.isNullOrEmpty(zone), "Zone must not be empty or null.");
     Preconditions.checkArgument(
         !Strings.isNullOrEmpty(cluster), "Cluster must not be empty or null.");
-    Preconditions.checkNotNull(channelOptions, "ChannelCoptions must not be empty or null.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(userAgent),
+        "UserAgent must not be empty or null");
+    Preconditions.checkArgument(channelCount > 0, "Channel count has to be at least 1.");
+    Preconditions.checkArgument(timeoutMs >= -1,
+      "ChannelTimeoutMs has to be positive, or -1 for none.");
+
     this.tableAdminHost = Preconditions.checkNotNull(tableAdminHost);
     this.clusterAdminHost = Preconditions.checkNotNull(clusterAdminHost);
     this.dataHost = Preconditions.checkNotNull(dataHost);
@@ -147,7 +223,14 @@ public class BigtableOptions {
     this.projectId = projectId;
     this.zone = zone;
     this.cluster = cluster;
-    this.channelOptions = channelOptions;
+    this.credential = credential;
+    this.authority = authority;
+    this.userAgent = userAgent;
+    this.callTimingReportPath = callTimingReportPath;
+    this.callStatusReportPath = callStatusReportPath;
+    this.retryOptions = retryOptions;
+    this.timeoutMs = timeoutMs;
+    this.channelCount = channelCount;
 
     LOG.debug("Connection Configuration: project: %s, cluster: %s, data host %s, "
         + "table admin host %s, cluster admin host %s using transport %s.",
@@ -171,10 +254,6 @@ public class BigtableOptions {
     return cluster;
   }
 
-  public ChannelOptions getChannelOptions() {
-    return channelOptions;
-  }
-
   public InetAddress getDataHost() {
     return dataHost;
   }
@@ -189,5 +268,62 @@ public class BigtableOptions {
 
   public int getPort() {
     return port;
+  }
+
+  /**
+   * Get the credential this object was constructed with. May be null.
+   * @return Null to indicate no credentials, otherwise, the Credentials object.
+   */
+  public Credentials getCredential() {
+    return credential;
+  }
+
+  /**
+   * Gets the authority to be passed in the HTTP/2 headers when creating new streams
+   * for the channel.
+   */
+  public String getAuthority() {
+    return authority;
+  }
+
+  /**
+   * Gets the user-agent to be appended to User-Agent header when creating new streams
+   * for the channel.
+   */
+  public String getUserAgent() {
+    return userAgent;
+  }
+
+  /**
+   * Get the client-local file to which a report of call timings will be appended.
+   */
+  public String getCallTimingReportPath() {
+    return callTimingReportPath;
+  }
+
+  /**
+   * Get the client-local file to which a report of call statuses will be appended.
+   */
+  public String getCallStatusReportPath() {
+    return callStatusReportPath;
+  }
+
+  /**
+   * Options controlling retries.
+   */
+  public RetryOptions getRetryOptions() { return retryOptions; }
+
+  /**
+   * The timeout for a channel.
+   */
+  public long getTimeoutMs() {
+    return timeoutMs;
+  }
+
+  /**
+   * The number of channels to create.
+   */
+  public int getChannelCount() {
+    return channelCount;
   }
 }
